moved sym_module to new module
