+++ tests/programs/conditionals/simpleWhile.oox
@@ -1,0 +1,12 @@
+
+class Main {
+    static void main(int x){
+        assume 1 < x;
+        x := x - 1;
+        assume 0 < x;
+        //assume 1 < x;
+        //while (0 < x) x := x - 1;
+
+    }
+}
+
